[FIX] error mng
