+++ delivery_roulier_laposte/models/stock.py
@@ -5,7 +5,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 import logging
-from openerp import models, fields, api, _
+from openerp import models, fields, api
 from openerp.tools import DEFAULT_SERVER_DATETIME_FORMAT
 
 from datetime import datetime, timedelta
@@ -72,7 +72,7 @@ class StockPicking(models.Model):
                 options['recommendationLevel'] = self.laposte_insur_recomm
                 del options['insuranceValue']
             else:
-                options['insuranceValue'] = self.laposte_insur_recomm
+                options['insuranceValue'] = int(self.laposte_insur_recomm)
         if 'cod' in options:
             options['codAmount'] = 0
         return options
@@ -113,46 +113,3 @@ class StockPicking(models.Model):
                 and partner.firstname:
             address['firstName'] = partner.firstname
         return address
-
-    @api.model
-    def _laposte_error_handling(self, payload, response):
-        ret_mess = u'Données transmises:\n%s\n\nExceptions levées%s\n%s'
-        if response.get('api_call_exception'):
-            # InvalidInputException
-            # on met des clés plus explicites vis à vis des objets odoo
-            suffix = (u"\nSignification des clés dans le contexte Odoo:\n"
-                      u"- 'to_address' correspond à 'adresse client'\n"
-                      u"- 'from_address' correspond à 'adresse de la société'")
-            password = payload['auth']['password']
-            message = u'Données transmises:\n%s\n\nExceptions levées%s\n%s' % (
-                payload, response, suffix)
-            message = message.replace(password, '****')
-            return message
-        elif response.get('message'):
-            # Webservice error
-            # on contextualise les réponses ws aux objets Odoo
-            map_responses = {
-                30204:
-                    u"La 2eme ligne d'adresse du client partenaire "
-                    u"est vide ou invalide",
-                30221:
-                    u"Le telephone du client ne doit comporter que des "
-                    u"chiffres ou le symbole +: convertissez tous vos N° de "
-                    u"telephone au format standard a partir du menu suivant:\n"
-                    u"Configuration > Technique > Telephonie > Reformate "
-                    u"les numeros de telephone ",
-                30100:
-                    u"La seconde ligne d'adresse de l'expéditeur est "
-                    u"vide ou invalide.",
-            }
-            message = response.get('message')
-            param_message = {'ws_exception': message,
-                             'resolution': ''}
-            if message and message.get('id') in map_responses.keys():
-                param_message['resolution'] = map_responses[
-                    message['id']]
-            ret_mess = _("Incident\n-----------\nReponse de Laposte:\n"
-                         "%(ws_exception)s\n\n"
-                         "Resolution\n-------------\n%(resolution)s"
-                         % param_message)
-        return ret_mess
